requirements/base.txt
asn1crypto==0.24.0
astroid==2.0.4
autopep8==1.5
Babel==2.9.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
bs4==0.0.1
boto3==1.15.17
certifi==2019.3.9
cffi==1.13.2
chardet==3.0.4
colorama==0.3.9
cryptography==3.4.8
deep-translator==1.3.2
EasyProcess
gspread==3.1.0
httplib2==0.13.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.6.3
mccabe==0.6.1
mysql-connector==2.2.9
oauth2client==4.1.3
paramiko==2.6.0
phpserialize==1.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pybase64==1.0.0
pycparser==2.19
pylint==2.1.1
pymongo==3.8.0
PyNaCl==1.4.0
python-dateutil==2.8.0
python-slugify==3.0.2
pytz==2019.2
PyVirtualDisplay
PyYAML==5.1
requests==2.21.0
rsa==4.0
selenium==3.141.0
selenium-wire==1.0.11
six==1.14.0
soupsieve==1.9.1
sshtunnel==0.1.5
text-unidecode==1.2
tinys3==0.1.12
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.24.2
wincertstore==0.2
wrapt==1.10.11
XlsxWriter==1.2.0
xmltodict==0.12.0
requests-html==0.10.0
sharepy==1.3.0
googletrans==3.0.0
rope==0.17.0
textblob==0.15.3
japanese-numbers-python==0.2.1
langdetect==1.0.8
word2number==1.1
pyotp
playwright
undetected-chromedriver
pycountry
webdriver-manager==3.4.2